replaced specific do repos with general do repo
