+++ apps/server/src/modules/board/board.module.ts
@@ -2,7 +2,7 @@ import { Module } from '@nestjs/common';
 import { ConsoleWriterModule } from '@shared/infra/console';
 import { LoggerModule } from '@src/core/logger';
 import { BoardManagementConsole } from './console';
-import { BoardNodeRepo, CardRepo, ColumnBoardRepo, ColumnRepo, ContentElementRepo } from './repo';
+import { BoardNodeRepo } from './repo';
 import { ColumnBoardService, CardService, ContentElementService } from './service';
 import { BoardManagementUc } from './uc';
 
@@ -11,10 +11,6 @@ import { BoardManagementUc } from './uc';
 	providers: [
 		BoardManagementConsole,
 		BoardManagementUc,
-		ColumnBoardRepo,
-		ColumnRepo,
-		CardRepo,
-		ContentElementRepo,
 		BoardNodeRepo,
 		ColumnBoardService,
 		CardService,
